[WIP] Resolve MSAA framebuffers on glBindFramebuffer() and glCopyTexImage2D()
Cogs works with MSAA after this. Note that there are several other GL
functions which would need shims to do MSAA resolves: all the
glReadPixels(), glCopyPixels(), glCopyTexSubImage2D(), etc, functions
which read from the framebuffer (and which implicitly resolve with the
default framebuffer, but just fail with an MSAA FBO).

And, of course, there are even more functions if we support
EXT_direct_state_access and friends...