Support SDL_PHYSPAL
Add a separate SDL_Palette to represent the physical palette, and swap
it in and out in PresentScreen().

Note that we have to manually swap the pointers in
surface20->format->palette, as SDL20_SetSurfacePalette seems to just
update the colours(?), resulting in the logical and physical palettes
ending up the same.