Support more possible formats for VideoTexture20
Currently, sdl12-compat always attempts to create the backing texture in
the first available texture format, and will convert to this format in
software (via the VideoConvertSurface20 pseudo-surface) if the
application's requested format is anything else.

Since most SDL2 renderers actually support a lot of different texture
formats, enumerate through all of them to find a compatible texture
format if one exists. This makes it much less likely to need to use the
fallback conversion path.