Super hacky "quirks" implementation
There are enough application-specific compatibility workarounds that it
may be worth keeping a database of them and automatically applying them.

This implementation adds a "quirks table": an array listing the
executable name and a series of string→string maps which match the
"hints" we've previously used manually.

All of those hints now go via an SDL12Compat_GetHint() function, which
will check the environment variable first, then look up a quirk based on
the executable name, then fallback to the default.

There are several flaws with this approach:
- Do we want to be in the business of per-application workarounds within
  sdl12-compat at all?
- Is the executable name a good way of distinguishing applications? Can
  we wait until they set a window title instead? Do we need wildcards to
  handle, e.g., ".bin.x86" and ".bin.x86_64" versions with the same
  hints.
- The "GetExeName" code is horrible, platform-specific, and largely
  untested. It works for a demo, though.
- The exact API implementation is kinda-horrible, but that at least can
  be fixed. :)

If you can stomach the xml, Mesa's driconf is a pretty robust
implementation of exactly this, and has some interesting features like
triggering workarounds based on hashes of the executable and other
things.

To act as a better demo, I also added a new
SDL12COMPAT_FORCE_GL_SWAPBUFFER_CONTEXT hint to make Awesomenauts work.
It's not an ideal fix — surely there must be some reason it works under
SDL 1.2 at all — but works okay here.

Thoughts?